✨ feat(prepare_words): add new filtering methods
- introduce filter_excludes_letters to filter words excluding specified letters
- introduce filter_includes_letters to filter words including only specified letters
- update tests to cover new filtering methods
